refactor(actions): refactor actions
