+++ packages/web-app/src/components/appli/EntitiesForm/Organization/index.js
@@ -21,8 +21,8 @@ import { useBoolean } from '../../../../hooks';
 import ActionButton from '../../../common/ActionButton';
 import Alert from '../../../common/Alert';
 import { updateName } from '../../../../actions/Name';
-import { postOrganization } from '../../../../actions/CreateOrganization';
-import { updateOrganization } from '../../../../actions/UpdateOrganization';
+import { postOrganization } from '../../../../actions/Organization/CreateOrganization';
+import { updateOrganization } from '../../../../actions/Organization/UpdateOrganization';
 import BasicInformationsForm from './BasicInformationsForm';
 import InformationsForm from './InformationsForm';
 import License from './License';
