+++ packages/web-app/src/reducers/OrganizationPutReducer.js
@@ -2,7 +2,7 @@ import {
   UPDATE_ORGANIZATION,
   UPDATE_ORGANIZATION_FAILURE,
   UPDATE_ORGANIZATION_SUCCESS
-} from '../actions/UpdateOrganization';
+} from '../actions/Organization/UpdateOrganization';
 
 const initialState = {
   loading: false,
